Remove unnecesary tests
