+++ tests/amm_contracts/amm.rs
@@ -1,17 +1,21 @@
+use std::collections::HashSet;
+
 use cairo_rs::vm::runners::cairo_runner::ExecutionResources;
 use felt::Felt;
 use num_traits::Zero;
 use starknet_rs::{
     business_logic::{
         execution::objects::{CallInfo, CallType},
-        fact_state::state::ExecutionResourcesManager,
+        fact_state::{
+            in_memory_state_reader::InMemoryStateReader, state::ExecutionResourcesManager,
+        },
+        state::cached_state::CachedState,
         transaction::error::TransactionError,
     },
     definitions::general_config::StarknetGeneralConfig,
     services::api::contract_class::EntryPointType,
     utils::{calculate_sn_keccak, Address},
 };
-use std::collections::HashSet;
 
 use crate::amm_contracts::utils::*;
 
@@ -20,6 +24,13 @@ fn init_pool(
     call_config: &mut CallConfig,
 ) -> Result<CallInfo, TransactionError> {
     execute_entry_point("init_pool", calldata, call_config)
+}
+
+fn add_demo_token(
+    calldata: &[Felt],
+    call_config: &mut CallConfig,
+) -> Result<CallInfo, TransactionError> {
+    execute_entry_point("add_demo_token", calldata, call_config)
 }
 
 fn get_pool_token_balance(
@@ -36,26 +47,20 @@ fn get_account_token_balance(
     execute_entry_point("get_account_token_balance", calldata, call_config)
 }
 
-fn add_demo_token(
-    calldata: &[Felt],
-    call_config: &mut CallConfig,
-) -> Result<CallInfo, TransactionError> {
-    execute_entry_point("add_demo_token", calldata, call_config)
-}
-
 fn swap(calldata: &[Felt], call_config: &mut CallConfig) -> Result<CallInfo, TransactionError> {
     execute_entry_point("swap", calldata, call_config)
 }
 
 #[test]
 fn amm_init_pool_test() {
-    let address = Address(1111.into());
-    let class_hash = [1; 32];
-    let mut state = setup_contract("starknet_programs/amm.json", &address, class_hash);
+    let general_config = StarknetGeneralConfig::default();
+    let mut state = CachedState::new(InMemoryStateReader::default(), Some(Default::default()));
+    // Deploy contract
+    let (contract_address, class_hash) =
+        deploy(&mut state, "starknet_programs/amm.json", &general_config);
 
     let calldata = [10000.into(), 10000.into()].to_vec();
     let caller_address = Address(0000.into());
-    let general_config = StarknetGeneralConfig::default();
     let mut resources_manager = ExecutionResourcesManager::default();
 
     let amm_entrypoint_selector = Felt::from_bytes_be(&calculate_sn_keccak(b"init_pool"));
@@ -66,7 +71,7 @@ fn amm_init_pool_test() {
     let expected_call_info = CallInfo {
         caller_address: Address(0.into()),
         call_type: Some(CallType::Delegate),
-        contract_address: Address(1111.into()),
+        contract_address: contract_address.clone(),
         entry_point_selector: Some(amm_entrypoint_selector),
         entry_point_type: Some(EntryPointType::External),
         calldata: calldata.clone(),
@@ -80,7 +85,7 @@ fn amm_init_pool_test() {
     let mut call_config = CallConfig {
         state: &mut state,
         caller_address: &caller_address,
-        address: &address,
+        address: &contract_address,
         class_hash: &class_hash,
         general_config: &general_config,
         resources_manager: &mut resources_manager,
@@ -94,27 +99,20 @@ fn amm_init_pool_test() {
 
 #[test]
 fn amm_add_demo_tokens_test() {
-    let address = Address(1111.into());
-    let class_hash = [1; 32];
-    let mut state = setup_contract("starknet_programs/amm.json", &address, class_hash);
+    let general_config = StarknetGeneralConfig::default();
+    let mut state = CachedState::new(InMemoryStateReader::default(), Some(Default::default()));
+    // Deploy contract
+    let (contract_address, class_hash) =
+        deploy(&mut state, "starknet_programs/amm.json", &general_config);
 
     let calldata = [10000.into(), 10000.into()].to_vec();
     let caller_address = Address(0000.into());
-    let general_config = StarknetGeneralConfig::default();
     let mut resources_manager = ExecutionResourcesManager::default();
-
-    let accessed_storage_keys_add_demo_token = get_accessed_keys(
-        "account_balance",
-        vec![
-            vec![0_u8.into(), 1_u8.into()],
-            vec![0_u8.into(), 2_u8.into()],
-        ],
-    );
 
     let mut call_config = CallConfig {
         state: &mut state,
         caller_address: &caller_address,
-        address: &address,
+        address: &contract_address,
         class_hash: &class_hash,
         general_config: &general_config,
         resources_manager: &mut resources_manager,
@@ -126,10 +124,18 @@ fn amm_add_demo_tokens_test() {
 
     let add_demo_token_selector = Felt::from_bytes_be(&calculate_sn_keccak(b"add_demo_token"));
 
+    let accessed_storage_keys_add_demo_token = get_accessed_keys(
+        "account_balance",
+        vec![
+            vec![0_u8.into(), 1_u8.into()],
+            vec![0_u8.into(), 2_u8.into()],
+        ],
+    );
+
     let expected_call_info_add_demo_token = CallInfo {
         caller_address: Address(0.into()),
         call_type: Some(CallType::Delegate),
-        contract_address: Address(1111.into()),
+        contract_address: contract_address.clone(),
         entry_point_selector: Some(add_demo_token_selector),
         entry_point_type: Some(EntryPointType::External),
         calldata: calldata_add_demo_token.clone(),
@@ -148,19 +154,20 @@ fn amm_add_demo_tokens_test() {
 
 #[test]
 fn amm_get_pool_token_balance() {
-    let address = Address(1111.into());
-    let class_hash = [1; 32];
-    let mut state = setup_contract("starknet_programs/amm.json", &address, class_hash);
+    let general_config = StarknetGeneralConfig::default();
+    let mut state = CachedState::new(InMemoryStateReader::default(), Some(Default::default()));
+    // Deploy contract
+    let (contract_address, class_hash) =
+        deploy(&mut state, "starknet_programs/amm.json", &general_config);
 
     let calldata = [10000.into(), 10000.into()].to_vec();
     let caller_address = Address(0000.into());
-    let general_config = StarknetGeneralConfig::default();
     let mut resources_manager = ExecutionResourcesManager::default();
 
     let mut call_config = CallConfig {
         state: &mut state,
         caller_address: &caller_address,
-        address: &address,
+        address: &contract_address,
         class_hash: &class_hash,
         general_config: &general_config,
         resources_manager: &mut resources_manager,
@@ -168,56 +175,51 @@ fn amm_get_pool_token_balance() {
 
     init_pool(&calldata, &mut call_config).unwrap();
 
-    let calldata_getter = [1.into()].to_vec();
+    let calldata_get_pool_token_balance = [1.into()].to_vec();
 
-    let get_pool_balance_selector =
+    let get_pool_token_balance_selector =
         Felt::from_bytes_be(&calculate_sn_keccak(b"get_pool_token_balance"));
 
-    let result = get_pool_token_balance(&calldata_getter, &mut call_config);
-
-    let accessed_storage_keys_pool_balance =
+    let accessed_storage_keys_get_pool_token_balance =
         get_accessed_keys("pool_balance", vec![vec![1_u8.into()]]);
 
-    let expected_call_info_getter = CallInfo {
+    let expected_call_info_get_pool_token_balance = CallInfo {
         caller_address: Address(0.into()),
         call_type: Some(CallType::Delegate),
-        contract_address: Address(1111.into()),
-        entry_point_selector: Some(get_pool_balance_selector),
+        contract_address: contract_address.clone(),
+        entry_point_selector: Some(get_pool_token_balance_selector),
         entry_point_type: Some(EntryPointType::External),
-        calldata: calldata_getter,
-        retdata: [10000.into()].to_vec(),
+        calldata: calldata_get_pool_token_balance.clone(),
         execution_resources: ExecutionResources::default(),
         class_hash: Some(class_hash),
-        accessed_storage_keys: accessed_storage_keys_pool_balance,
-        storage_read_values: [10000.into()].to_vec(),
+        accessed_storage_keys: accessed_storage_keys_get_pool_token_balance,
+        storage_read_values: vec![10000.into()],
+        retdata: [10000.into()].to_vec(),
         ..Default::default()
     };
 
-    assert_eq!(result.unwrap(), expected_call_info_getter);
+    assert_eq!(
+        get_pool_token_balance(&calldata_get_pool_token_balance, &mut call_config).unwrap(),
+        expected_call_info_get_pool_token_balance
+    );
 }
 
-//test swap functionality
-//first we set up the contract
-//second we init the pool
-//third we add tokens to the user
-//fourth we swap tokens
-//check if its what we expected
 #[test]
 fn amm_swap_test() {
-    //set up contract
-    let address = Address(1111.into());
-    let class_hash = [1; 32];
-    let mut state = setup_contract("starknet_programs/amm.json", &address, class_hash);
+    let general_config = StarknetGeneralConfig::default();
+    let mut state = CachedState::new(InMemoryStateReader::default(), Some(Default::default()));
+    // Deploy contract
+    let (contract_address, class_hash) =
+        deploy(&mut state, "starknet_programs/amm.json", &general_config);
 
     let calldata = [10000.into(), 10000.into()].to_vec();
     let caller_address = Address(0000.into());
-    let general_config = StarknetGeneralConfig::default();
     let mut resources_manager = ExecutionResourcesManager::default();
 
     let mut call_config = CallConfig {
         state: &mut state,
         caller_address: &caller_address,
-        address: &address,
+        address: &contract_address,
         class_hash: &class_hash,
         general_config: &general_config,
         resources_manager: &mut resources_manager,
@@ -229,13 +231,11 @@ fn amm_swap_test() {
     let calldata_add_demo_token = [100.into(), 100.into()].to_vec();
     add_demo_token(&calldata_add_demo_token, &mut call_config).unwrap();
 
-    //swap tokens. Token 1 with 10 in amount
-    let calldataswap = [1.into(), 10.into()].to_vec();
+    let calldata_swap = [1.into(), 10.into()].to_vec();
 
-    //expected return value 9
     let expected_return = [9.into()].to_vec();
 
-    let result = swap(&calldataswap, &mut call_config);
+    let swap_selector = Felt::from_bytes_be(&calculate_sn_keccak(b"swap"));
 
     //access keys are all keys in pool balance and only this users balance but thats checked in account
     let accessed_storage_keys_pool_balance =
@@ -254,15 +254,13 @@ fn amm_swap_test() {
     accessed_storage_keys.extend(accessed_storage_keys_pool_balance);
     accessed_storage_keys.extend(accessed_storage_keys_user_balance);
 
-    let swap_selector = Felt::from_bytes_be(&calculate_sn_keccak(b"swap"));
-
-    let expected_call_infoswap = CallInfo {
+    let expected_call_info_swap = CallInfo {
         caller_address: Address(0.into()),
         call_type: Some(CallType::Delegate),
-        contract_address: Address(1111.into()),
+        contract_address: contract_address.clone(),
         entry_point_selector: Some(swap_selector),
         entry_point_type: Some(EntryPointType::External),
-        calldata: calldataswap,
+        calldata: calldata_swap.clone(),
         retdata: expected_return,
         execution_resources: ExecutionResources::default(),
         class_hash: Some(class_hash),
@@ -278,14 +276,19 @@ fn amm_swap_test() {
         ..Default::default()
     };
 
-    assert_eq!(result.unwrap(), expected_call_infoswap);
+    assert_eq!(
+        swap(&calldata_swap, &mut call_config).unwrap(),
+        expected_call_info_swap
+    );
 }
 
 #[test]
 fn amm_init_pool_should_fail_with_amount_out_of_bounds() {
-    let address = Address(1111.into());
-    let class_hash = [1; 32];
-    let mut state = setup_contract("starknet_programs/amm.json", &address, class_hash);
+    let general_config = StarknetGeneralConfig::default();
+    let mut state = CachedState::new(InMemoryStateReader::default(), Some(Default::default()));
+    // Deploy contract
+    let (contract_address, class_hash) =
+        deploy(&mut state, "starknet_programs/amm.json", &general_config);
 
     let calldata = [Felt::new(2_u32.pow(30)), Felt::new(2_u32.pow(30))].to_vec();
     let caller_address = Address(0000.into());
@@ -294,7 +297,7 @@ fn amm_init_pool_should_fail_with_amount_out_of_bounds() {
     let mut call_config = CallConfig {
         state: &mut state,
         caller_address: &caller_address,
-        address: &address,
+        address: &contract_address,
         class_hash: &class_hash,
         general_config: &general_config,
         resources_manager: &mut resources_manager,
@@ -305,9 +308,11 @@ fn amm_init_pool_should_fail_with_amount_out_of_bounds() {
 
 #[test]
 fn amm_swap_should_fail_with_unexistent_token() {
-    let address = Address(1111.into());
-    let class_hash = [1; 32];
-    let mut state = setup_contract("starknet_programs/amm.json", &address, class_hash);
+    let general_config = StarknetGeneralConfig::default();
+    let mut state = CachedState::new(InMemoryStateReader::default(), Some(Default::default()));
+    // Deploy contract
+    let (contract_address, class_hash) =
+        deploy(&mut state, "starknet_programs/amm.json", &general_config);
 
     let calldata = [Felt::zero(), Felt::new(10)].to_vec();
     let caller_address = Address(0000.into());
@@ -316,7 +321,7 @@ fn amm_swap_should_fail_with_unexistent_token() {
     let mut call_config = CallConfig {
         state: &mut state,
         caller_address: &caller_address,
-        address: &address,
+        address: &contract_address,
         class_hash: &class_hash,
         general_config: &general_config,
         resources_manager: &mut resources_manager,
@@ -327,9 +332,11 @@ fn amm_swap_should_fail_with_unexistent_token() {
 
 #[test]
 fn amm_swap_should_fail_with_amount_out_of_bounds() {
-    let address = Address(1111.into());
-    let class_hash = [1; 32];
-    let mut state = setup_contract("starknet_programs/amm.json", &address, class_hash);
+    let general_config = StarknetGeneralConfig::default();
+    let mut state = CachedState::new(InMemoryStateReader::default(), Some(Default::default()));
+    // Deploy contract
+    let (contract_address, class_hash) =
+        deploy(&mut state, "starknet_programs/amm.json", &general_config);
 
     let calldata = [Felt::new(1), Felt::new(2_u32.pow(30))].to_vec();
     let caller_address = Address(0000.into());
@@ -338,7 +345,7 @@ fn amm_swap_should_fail_with_amount_out_of_bounds() {
     let mut call_config = CallConfig {
         state: &mut state,
         caller_address: &caller_address,
-        address: &address,
+        address: &contract_address,
         class_hash: &class_hash,
         general_config: &general_config,
         resources_manager: &mut resources_manager,
@@ -349,9 +356,12 @@ fn amm_swap_should_fail_with_amount_out_of_bounds() {
 
 #[test]
 fn amm_swap_should_fail_when_user_does_not_have_enough_funds() {
-    let address = Address(1111.into());
-    let class_hash = [1; 32];
-    let mut state = setup_contract("starknet_programs/amm.json", &address, class_hash);
+    let general_config = StarknetGeneralConfig::default();
+    let mut state = CachedState::new(InMemoryStateReader::default(), Some(Default::default()));
+    // Deploy contract
+    let (contract_address, class_hash) =
+        deploy(&mut state, "starknet_programs/amm.json", &general_config);
+
     let calldata = [Felt::new(1), Felt::new(100)].to_vec();
     let caller_address = Address(0000.into());
     let general_config = StarknetGeneralConfig::default();
@@ -359,7 +369,7 @@ fn amm_swap_should_fail_when_user_does_not_have_enough_funds() {
     let mut call_config = CallConfig {
         state: &mut state,
         caller_address: &caller_address,
-        address: &address,
+        address: &contract_address,
         class_hash: &class_hash,
         general_config: &general_config,
         resources_manager: &mut resources_manager,
@@ -373,9 +383,11 @@ fn amm_swap_should_fail_when_user_does_not_have_enough_funds() {
 
 #[test]
 fn amm_get_account_token_balance_test() {
-    let address = Address(1111.into());
-    let class_hash = [1; 32];
-    let mut state = setup_contract("starknet_programs/amm.json", &address, class_hash);
+    let general_config = StarknetGeneralConfig::default();
+    let mut state = CachedState::new(InMemoryStateReader::default(), Some(Default::default()));
+    // Deploy contract
+    let (contract_address, class_hash) =
+        deploy(&mut state, "starknet_programs/amm.json", &general_config);
 
     //add 10 tokens of token type 1
     let caller_address = Address(0000.into());
@@ -386,7 +398,7 @@ fn amm_get_account_token_balance_test() {
     let mut call_config = CallConfig {
         state: &mut state,
         caller_address: &caller_address,
-        address: &address,
+        address: &contract_address,
         class_hash: &class_hash,
         general_config: &general_config,
         resources_manager: &mut resources_manager,
@@ -409,7 +421,7 @@ fn amm_get_account_token_balance_test() {
     let expected_call_info_get_account_token_balance = CallInfo {
         caller_address: Address(0.into()),
         call_type: Some(CallType::Delegate),
-        contract_address: Address(1111.into()),
+        contract_address,
         entry_point_selector: Some(get_account_token_balance_selector),
         entry_point_type: Some(EntryPointType::External),
         calldata: calldata_get_balance,
