+++ tests/amm_contracts/mod.rs
@@ -1,5 +1,3 @@
 pub mod amm;
-pub mod amm_deploy;
 pub mod amm_proxy;
-pub mod amm_proxy_deploy;
 pub mod utils;
